+++ CHANGELOG.md
@@ -5,7 +5,11 @@ Changelog for insanic
 0.4.10 (unreleased)
 -------------------
 
-- Nothing changed yet.
+- FEAT: `files` parameter on `http_dispatch` to send files. 
+    - Will place `multipart/form-data` as content-type
+- FEAT: `request_timeout` on `http_dispatch` to override default
+- FIX: register target as ip instead of hostname on kong
+
 
 
 0.4.9 (2018-06-22)